Use the context to decide whether ModSecurity is enabled or not
Currently, the connector decides whether ModSecurity is enabled or not
based on the configuration directive in the request's location conf.
In case of an internal redirect and if ModSecurity is not enabled in the
internal redirect's location conf, the log handler will not run and the
transaction of the original request will not be logged.

This commit extends the current behavior of the header and body filters,
which assume that a null context means that ModSecurity is disabled, to
the pre access and log handlers. As a result, the connector will always
decide based on the original request whether ModSecurity is enabled or not.